單筆: 整理程式碼(把redis相關功能，集中在RedisService)
