+++ src/main/java/com/example/pentaho/service/RedisService.java
@@ -1,0 +1,123 @@
+package com.example.pentaho.service;
+
+import com.example.pentaho.component.SingleQueryDTO;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.redis.core.*;
+import org.springframework.stereotype.Service;
+
+import java.util.*;
+import java.util.concurrent.ConcurrentSkipListSet;
+
+@Service
+public class RedisService {
+
+
+    private static Logger log = LoggerFactory.getLogger(SingleQueryService.class);
+
+    Integer SCAN_SIZE = 1000;
+
+    @Autowired
+    private StringRedisTemplate stringRedisTemplate;
+
+
+    /**
+     * 找為LIST的值 (redis: LRANGE)
+     */
+    public List<String> findListByKey(String key) {
+        ListOperations<String, String> listOps = stringRedisTemplate.opsForList();
+        List<String> elements = listOps.range(key, 0, -1);
+        log.info("elements:{}", elements);
+        return elements;
+    }
+
+    /**
+     * set單一個值 (redis: set)
+     */
+    public void setData(SingleQueryDTO singleQueryDTO) {
+        stringRedisTemplate.opsForValue().set(singleQueryDTO.getRedisKey(), singleQueryDTO.getRedisValue());
+        log.info("set單一個值，key: {}, value: {}", singleQueryDTO.getRedisKey(), singleQueryDTO.getRedisValue());
+    }
+
+
+    /**
+     * 一個key塞多筆值(redis: RPUSH)
+     */
+    public void pushData(SingleQueryDTO singleQueryDTO) {
+        stringRedisTemplate.opsForList().rightPushAll(singleQueryDTO.getRedisKey(), singleQueryDTO.getRedisValueList());
+        log.info("push value to a key，key: {}, value: {}", singleQueryDTO.getRedisKey(), singleQueryDTO.getRedisValueList());
+    }
+
+
+    /**
+     * 找為SET的值 (redis: SMEMBERS)
+     */
+    public Set<String> findSetByKey(SingleQueryDTO singleQueryDTO) {
+        SetOperations<String, String> setOps = stringRedisTemplate.opsForSet();
+        Set<String> elements = setOps.members(singleQueryDTO.getRedisKey());
+        log.info("elements:{}", elements);
+        return elements;
+    }
+
+
+    /**
+     * 找單一個值 (redis: get)
+     */
+    public String findByKey(String columnName, String key, String defaultValue) {
+        if (key != null && !key.isEmpty()) {
+            String redisValue = stringRedisTemplate.opsForValue().get(key);
+            if (redisValue != null && !redisValue.isEmpty()) {
+                log.info("columnName:{} , redisKey: {} , redisValue: {}", columnName, key, redisValue);
+                return redisValue;
+            }
+        }
+        return defaultValue;
+    }
+
+
+    public Map<String, String> findByKeys(Map<String, String> keyMap, String segmentExistNumber) {
+        Map<String, String> resultMap = new HashMap<>();
+        List<String> redisKeys = new ArrayList<>(keyMap.keySet());
+        List<String> redisValues = stringRedisTemplate.opsForValue().multiGet(redisKeys);
+        for (int i = 0; i < redisKeys.size(); i++) {
+            String key = redisKeys.get(i);
+            String redisValue = redisValues.get(i);
+            if (redisValue != null && !redisValue.isEmpty()) {
+                log.info("redisKey: {} , redisValue: {}", key, redisValue);
+                resultMap.put(key, redisValue);
+                segmentExistNumber += "1";
+            } else {
+                log.info("redisKey: {} , 找不到redisValue: {}", key, redisValue);
+                resultMap.put(key, keyMap.get(key)); //如果找不到對應的value的話，就要放default value
+                segmentExistNumber += "0";
+            }
+        }
+        resultMap.put("segmentExistNumber", segmentExistNumber);
+        return resultMap;
+    }
+
+
+    /**
+     * 模糊比對，找出相符的 KEY (redis: scan)
+     */
+    public Set<String> findListByScan(String key) {
+        Set<String> keySet = stringRedisTemplate.execute((RedisCallback<Set<String>>) connection -> {
+            Set<String> keySetTemp = new ConcurrentSkipListSet<>();
+            try (Cursor<byte[]> cursor = connection.scan(ScanOptions.scanOptions()
+                    .match(key) //模糊比對
+                    .count(SCAN_SIZE)
+                    .build())) {
+                while (cursor.hasNext() && keySetTemp.size() < SCAN_SIZE) {
+                    keySetTemp.add(new String(cursor.next(), "utf-8"));
+                }
+            } catch (Exception e) {
+                log.error("redis，模糊比對錯誤:{}", e.getMessage());
+            }
+            return keySetTemp;
+        });
+        log.info("keySet:{}", keySet);
+        return keySet;
+    }
+
+}
